remove testing boilerplate, debug version of app
