+++ backend/src/authentication/endpoints.js
@@ -61,11 +61,3 @@ export function logout(req, res, next) {
     }
     res.status(401).send("Not logged in")
 }
-
-export function auth_test(req, res, next) {
-    if (req.isAuthenticated()) {
-        res.status(200).send(req.user.username)
-    } else {
-        res.status(401).send("Not authenticated")
-    }
-}
