+++ backend/src/debug/test-endpoints.js
@@ -1,0 +1,7 @@
+export function auth_test(req, res, next) {
+    if (req.isAuthenticated()) {
+        res.status(200).send(req.user.username)
+    } else {
+        res.status(401).send("Not authenticated")
+    }
+}
