+++ backend/tests/integration/authentication/endpoints.test.js
@@ -1,33 +1,16 @@
 import request from "supertest"
-import { MongoMemoryServer } from 'mongodb-memory-server';
-import { mongoose } from 'mongoose';
 
-import createApp from "../../../src/app.js";
 import { User } from "../../../src/authentication/schemas.js";
 import { hash } from "../../../src/authentication/utils.js";
 
-describe("POST /create_user", () => {
-    let mongoServer;
-    let con;
-  
-    beforeEach(async () => {
-        mongoServer = await MongoMemoryServer.create();
-        con = await mongoose.connect(mongoServer.getUri(), {});
-        process.env.MONGO_URL = mongoServer.getUri()
-    });
-  
-    afterEach(async () => {
-      await mongoose.disconnect()
-      if (con) {
-        await con.disconnect();
-      }
-      if (mongoServer) {
-        await mongoServer.stop();
-      }
-    });  
+import createTestApp from "../../../src/debug/test-app.js";
+import useMongoTestWrapper from "../../../src/debug/jest-mongo.js";
 
+describe("POST /create_user", () => {
+    useMongoTestWrapper() 
+ 
     it("should 400 without username or password", async () => {
-        const app = await createApp()
+        const app = await createTestApp()
 
         let response = await request(app).post("/create_user").send({})
         expect(response.statusCode).toBe(400);
@@ -44,7 +27,7 @@ describe("POST /create_user", () => {
         const password = "password"
 
         await (new User({username, password_hash: hash(password)})).save()
-        const response = await request(await createApp()).post("/create_user").send({username, password})
+        const response = await request(await createTestApp()).post("/create_user").send({username, password})
         expect(response.statusCode).toBe(409);
     })
 
@@ -52,7 +35,7 @@ describe("POST /create_user", () => {
         const username = "username"
         const password = "password"
 
-        const response = await request(await createApp()).post("/create_user").send({username, password})
+        const response = await request(await createTestApp()).post("/create_user").send({username, password})
         expect(response.statusCode).toBe(200);
         expect(response.body.username).toBe(username)
     })
@@ -60,7 +43,7 @@ describe("POST /create_user", () => {
     it("should be authenticated on success", async () => {
         const username = "username"
         const password = "password"
-        const app = await createApp()
+        const app = await createTestApp()
 
         let response = await request(app).get("/auth_test")
         expect(response.statusCode).toBe(401)
@@ -77,27 +60,10 @@ describe("POST /create_user", () => {
 })
 
 describe("POST /login", () => {
-    let mongoServer;
-    let con;
-  
-    beforeEach(async () => {
-        mongoServer = await MongoMemoryServer.create();
-        con = await mongoose.connect(mongoServer.getUri(), {});
-        process.env.MONGO_URL = mongoServer.getUri()
-    });
-  
-    afterEach(async () => {
-      await mongoose.disconnect()
-      if (con) {
-        await con.disconnect();
-      }
-      if (mongoServer) {
-        await mongoServer.stop();
-      }
-    });  
+    useMongoTestWrapper() 
 
     it("should 400 without username or password", async () => {
-        const app = await createApp()
+        const app = await createTestApp()
 
         let response = await request(app).post("/login").send({})
         expect(response.statusCode).toBe(400);
@@ -110,7 +76,7 @@ describe("POST /login", () => {
     })
 
     it("should 401 when user not found", async () => {
-        const response = await request(await createApp()).post("/login").send({username: "no", password: "nope"})
+        const response = await request(await createTestApp()).post("/login").send({username: "no", password: "nope"})
         expect(response.statusCode).toBe(401);
     })
 
@@ -120,7 +86,7 @@ describe("POST /login", () => {
         const not_password = "boohoo"
 
         await (new User({username, password_hash: hash(password)})).save()
-        const response = await request(await createApp()).post("/login").send({username, password: not_password})
+        const response = await request(await createTestApp()).post("/login").send({username, password: not_password})
         expect(response.statusCode).toBe(401);
     })
 
@@ -129,7 +95,7 @@ describe("POST /login", () => {
         const password = "password"
 
         await (new User({username, password_hash: hash(password)})).save()
-        const response = await request(await createApp()).post("/login").send({username, password})
+        const response = await request(await createTestApp()).post("/login").send({username, password})
         expect(response.statusCode).toBe(200);
         expect(response.body.username).toBe(username)
     })
@@ -137,7 +103,7 @@ describe("POST /login", () => {
     it("should be authenticated on success", async () => {
         const username = "username"
         const password = "password"
-        const app = await createApp()
+        const app = await createTestApp()
 
         let response = await request(app).get("/auth_test")
         expect(response.statusCode).toBe(401)
@@ -155,29 +121,12 @@ describe("POST /login", () => {
 })
 
 describe("DELETE /logout", () => {
-    let mongoServer;
-    let con;
-  
-    beforeEach(async () => {
-        mongoServer = await MongoMemoryServer.create();
-        con = await mongoose.connect(mongoServer.getUri(), {});
-        process.env.MONGO_URL = mongoServer.getUri()
-    });
-  
-    afterEach(async () => {
-      await mongoose.disconnect()
-      if (con) {
-        await con.disconnect();
-      }
-      if (mongoServer) {
-        await mongoServer.stop();
-      }
-    });
+    useMongoTestWrapper() 
 
     it("should 401 when not authenticated", async () => {
         const username = "username"
         const password = "password"
-        const app = await createApp()
+        const app = await createTestApp()
 
         let response = await request(app).delete("/logout")
         expect(response.statusCode).toBe(401)
@@ -186,7 +135,7 @@ describe("DELETE /logout", () => {
     it("should logout when authenticated", async () => {
         const username = "username"
         const password = "password"
-        const app = await createApp()
+        const app = await createTestApp()
 
         let response = await request(app).get("/auth_test")
         expect(response.statusCode).toBe(401)
