+++ backend/tests/unit/index.test.js
@@ -1,30 +1,16 @@
 import { mongoose } from 'mongoose';
-import { MongoMemoryServer } from 'mongodb-memory-server';
+
+import useMongoTestWrapper from '../../src/debug/jest-mongo';
 
 describe("Unit Testing Environment", () => {
-    let con;
-    let mongoServer;
-  
-    beforeEach(async () => {
-      mongoServer = await MongoMemoryServer.create();
-      con = await mongoose.connect(mongoServer.getUri(), {});
-    });
-  
-    afterEach(async () => {
-      if (con) {
-        await con.disconnect();
-      }
-      if (mongoServer) {
-        await mongoServer.stop();
-      }
-    });  
+    useMongoTestWrapper()  
 
     it("should run tests", () => {
         expect(3 + 5).toBe(8);
     })
 
     it('should use a test database', async () => {
-        const state = con.connection.readyState;
+        const state = mongoose.connection.readyState;
         const connectedState = 1
     
         expect(state).toBe(connectedState);
